Refactor constitution
