+++ src/felupe/constitution/_models_linear_elasticity.py
@@ -19,7 +19,7 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 import numpy as np
 
 from ..math import cdya, dya, identity, trace, transpose
-from ._preview import ConstitutiveMaterial
+from ._base import ConstitutiveMaterial
 
 
 def lame_converter(E, nu):
