+++ src/felupe/constitution/_user_materials.py
@@ -19,8 +19,8 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 import numpy as np
 
 from ..math import identity, ravel, reshape, sym
+from ._base import ConstitutiveMaterial
 from ._models_linear_elasticity import lame_converter
-from ._preview import ConstitutiveMaterial
 from ._user_materials_models import linear_elastic_plastic_isotropic_hardening
