+++ src/felupe/constitution/_view.py
@@ -483,166 +483,3 @@ class ViewMaterialIncompressible(PlotMaterial):
             P, statevars = self.umat.gradient([F, None])
 
         return λ1, (P[0, 0] - λ3 / λ1 * P[2, 2]).ravel(), "Biaxial (Incompressible)"
-
-
-class ConstitutiveMaterial:
-    "Base class for constitutive materials."
-
-    def view(self, incompressible=False, **kwargs):
-        """Create views on normal force per undeformed area vs. stretch curves for the
-        elementary homogeneous deformations uniaxial tension/compression, planar shear
-        and biaxial tension of a given isotropic material formulation.
-
-        Parameters
-        ----------
-        incompressible : bool, optional
-            A flag to enforce views on incompressible deformations (default is False).
-        **kwargs : dict, optional
-            Optional keyword-arguments for :class:`~felupe.ViewMaterial` or
-            :class:`~felupe.ViewMaterialIncompressible`.
-
-        Returns
-        -------
-        felupe.ViewMaterial or felupe.ViewMaterialIncompressible
-
-        See Also
-        --------
-        felupe.ViewMaterial : Create views on normal force per undeformed area vs.
-            stretch curves for the elementary homogeneous deformations uniaxial
-            tension/compression, planar shear and biaxial tension of a given isotropic
-            material formulation.
-        felupe.ViewMaterialIncompressible : Create views on normal force per undeformed
-            area vs. stretch curves for the elementary homogeneous incompressible
-            deformations uniaxial tension/compression, planar shear and biaxial tension
-            of a given isotropic material formulation.
-        """
-
-        View = ViewMaterial
-        if incompressible:
-            View = ViewMaterialIncompressible
-
-        return View(self, **kwargs)
-
-    def plot(self, incompressible=False, **kwargs):
-        """Return a plot with normal force per undeformed area vs. stretch curves for
-        the elementary homogeneous deformations uniaxial tension/compression, planar
-        shear and biaxial tension of a given isotropic material formulation.
-
-        Parameters
-        ----------
-        incompressible : bool, optional
-            A flag to enforce views on incompressible deformations (default is False).
-        **kwargs : dict, optional
-            Optional keyword-arguments for :class:`~felupe.ViewMaterial` or
-            :class:`~felupe.ViewMaterialIncompressible`.
-
-        Returns
-        -------
-        matplotlib.axes.Axes
-
-        See Also
-        --------
-        felupe.ViewMaterial : Create views on normal force per undeformed area vs.
-            stretch curves for the elementary homogeneous deformations uniaxial
-            tension/compression, planar shear and biaxial tension of a given isotropic
-            material formulation.
-        felupe.ViewMaterialIncompressible : Create views on normal force per undeformed
-            area vs. stretch curves for the elementary homogeneous incompressible
-            deformations uniaxial tension/compression, planar shear and biaxial tension
-            of a given isotropic material formulation.
-        """
-
-        return self.view(incompressible=incompressible, **kwargs).plot()
-
-    def screenshot(self, filename="umat.png", incompressible=False, **kwargs):
-        """Save a screenshot with normal force per undeformed area vs. stretch curves
-        for the elementary homogeneous deformations uniaxial tension/compression, planar
-        shear and biaxial tension of a given isotropic material formulation.
-
-        Parameters
-        ----------
-        filename : str, optional
-            The filename of the screenshot (default is "umat.png").
-        incompressible : bool, optional
-            A flag to enforce views on incompressible deformations (default is False).
-        **kwargs : dict, optional
-            Optional keyword-arguments for :class:`~felupe.ViewMaterial` or
-            :class:`~felupe.ViewMaterialIncompressible`.
-
-        Returns
-        -------
-        matplotlib.axes.Axes
-
-        See Also
-        --------
-        felupe.ViewMaterial : Create views on normal force per undeformed area vs.
-            stretch curves for the elementary homogeneous deformations uniaxial
-            tension/compression, planar shear and biaxial tension of a given isotropic
-            material formulation.
-        felupe.ViewMaterialIncompressible : Create views on normal force per undeformed
-            area vs. stretch curves for the elementary homogeneous incompressible
-            deformations uniaxial tension/compression, planar shear and biaxial tension
-            of a given isotropic material formulation.
-        """
-
-        import matplotlib.pyplot as plt
-
-        ax = self.plot(incompressible=incompressible, **kwargs)
-        fig = ax.get_figure()
-        fig.savefig(filename)
-        plt.close(fig)
-
-        return ax
-
-    def __and__(self, other_material):
-        return CompositeMaterial(self, other_material)
-
-
-class CompositeMaterial(ConstitutiveMaterial):
-    """A composite material with two constitutive materials merged. State variables are
-    only considered for the first material.
-
-    Parameters
-    ----------
-    material : ConstitutiveMaterial
-        First constitutive material.
-    other_material : ConstitutiveMaterial
-        Second constitutive material.
-
-    Notes
-    -----
-    ..  warning::
-        Do not merge two constitutive materials with the same keys of material
-        parameters. In this case, the values of these material parameters are taken from
-        the first constitutive material.
-
-    Examples
-    --------
-    >>> import felupe as fem
-    >>>
-    >>> nh = fem.NeoHooke(mu=1.0)
-    >>> vol = fem.Volumetric(bulk=2.0)
-    >>> umat = nh & vol
-    >>> ax = umat.plot()
-
-    ..  image:: images/umat_composite.png
-        :width: 400px
-
-    """
-
-    def __init__(self, material, other_material):
-        self.materials = [material, other_material]
-        self.kwargs = {**other_material.kwargs, **material.kwargs}
-        self.x = material.x
-
-    def gradient(self, x, **kwargs):
-        gradients = [material.gradient(x, **kwargs) for material in self.materials]
-        nfields = len(x) - 1
-        P = [np.sum([grad[i] for grad in gradients], axis=0) for i in range(nfields)]
-        statevars_new = gradients[0][-1]
-        return [*P, statevars_new]
-
-    def hessian(self, x, **kwargs):
-        hessians = [material.hessian(x, **kwargs) for material in self.materials]
-        nfields = len(x) - 1
-        return [np.sum([hess[i] for hess in hessians], axis=0) for i in range(nfields)]
